+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 4.0.0
+
+-  **New:** `variant` can have more than one value. Parsing the `variant` is too complex for this function, so any unknown tokens are assumed to be a `variant` and joined as one (e.g., `foo bold bar italic qux` parses a `variant` of `foo bar qux`).
+-  **Fix:** Fixed `normal` overriding other declarations. The defaults are still `normal` and you can have any number of normals precede the `size` without overriding non-`normal` values.
+-  **Breaking:** Removed parsing of CSS global keywords, as they don't seem to be supported. This is a fix, but it may break implementations that were relying on it.
+-  **Breaking:** Error messaging has changed and new errors will throw when adding more than one `style`, `weight` or `stretch`.
+
 ## 3.0.4
 
 -  Add BundlePhobia badges to npm (requires version bump).